test for assignment gql query; refactored test_helpers a bit; tests are less flakey
